Revert "Update README.md"

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,47 +1,36 @@
 package ru.yandex.practicum.filmorate.service;
 
 import lombok.Data;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.storage.FilmStorage;
-import ru.yandex.practicum.filmorate.storage.UserStorage;
+import ru.yandex.practicum.filmorate.storage.InMemoryFilmStorage;
+import ru.yandex.practicum.filmorate.storage.InMemoryUserStorage;
 
-import java.util.*;
+import java.util.Collection;
+import java.util.Comparator;
+import java.util.List;
 
 import static java.util.Objects.isNull;
 
 @Service
 @Slf4j
+@RequiredArgsConstructor
 @Data
-@Qualifier("newFilmStorage")
 public class FilmService {
 
-    private final FilmStorage filmStorage;
-    private final UserStorage userStorage;
-    private final MpaService mpaService;
-    private final GenreService genreService;
+    private final InMemoryFilmStorage filmStorage;
+    private final InMemoryUserStorage userStorage;
 
-    FilmService(@Qualifier("newFilmStorage") FilmStorage filmStorage, @Qualifier("newUserStorage") UserStorage userStorage, MpaService mpaService, GenreService genreService) {
-        this.filmStorage = filmStorage;
-        this.userStorage = userStorage;
-        this.mpaService = mpaService;
-        this.genreService = genreService;
-    }
-
-    public Optional<Film> findFilmById(Long id) {
+    public Film findFilmById(Long id) {
         return filmStorage.getFilm(id);
     }
 
     public Film createFilm(Film film) {
-        try {
-            mpaService.findMpaById(film.getMpa().getId());
-            return filmStorage.create(film);
-        } catch (NotFoundException notFoundException) {
-            throw new NotFoundException("Mpa не найден");
-        }
+        return filmStorage.create(film);
     }
 
     public Collection<Film> findAllFilm() {
@@ -61,14 +50,29 @@ public class FilmService {
         if (isNull(userStorage.getUser(userId))) {
             throw new NotFoundException("Пользователь не найден");
         }
+        if (filmStorage.findLikeFromUser(filmId, userId)) {
+            throw new ValidationException("Пользователем " + userStorage.getUser(userId).getName() + "лайки уже поставлен");
+        }
         filmStorage.addLikeFromFilm(filmId, userId);
     }
 
     public void deleteLikeFromFilm(Long filmId, Long userId) {
+        if (isNull(filmStorage.getFilm(filmId))) {
+            throw new NotFoundException("Указанного фильма не найдено");
+        }
+        if (isNull(userStorage.getUser(userId))) {
+            throw new NotFoundException("Указанного пользователя не найдено");
+        }
+        if (!filmStorage.getFilmMap().get(filmId).getLikeList().contains(userId)) {
+            throw new NotFoundException("Указанный пользователь лайк к фильму не ставил");
+        }
         filmStorage.deleteLikeFromFilm(filmId, userId);
     }
 
     public List<Film> favoriteFilm(Long x) {
-        return filmStorage.favoriteFilm(x);
+        return filmStorage.getFilmMap().values().stream()
+                .sorted(Comparator.comparingInt(film -> film.getLikeList().size()))
+                .limit(x)
+                .toList().reversed();
     }
 }
